data-migration-dq-checks/requirements.txt
aiohappyeyeballs==2.6.1
Bottleneck==1.3.7
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
clickhouse-connect==0.8.15
contourpy==1.3.1
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.56.0
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
httplib2==0.22.0
idna==3.10
kiwisolver==1.4.8
lz4==4.4.3
matplotlib==3.10.1
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.1.5
packaging==23.2
pandas==2.2.3
pillow==11.1.0
polars==1.26.0
proto-plus==1.26.1
protobuf==6.30.1
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
seaborn==0.13.2
six==1.16.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.3.0
zstandard==0.23.0
